+++ micro-ui/web/micro-ui-internals/packages/modules/workbench/src/Module.js
@@ -1,8 +1,6 @@
 import { TourProvider } from "@egovernments/digit-ui-react-components";
 import React from "react";
 import { useMatch } from "react-router-dom";
-import { Loader } from "@egovernments/digit-ui-components";
-
 import EmployeeApp from "./pages/employee";
 import { CustomisedHooks } from "./hooks";
 import { UICustomizations } from "./configs/UICustomizations";
@@ -10,15 +8,15 @@ import HRMSCard from "./components/HRMSCard";
 import WorkbenchCard from "./components/WorkbenchCard";
 import DigitJSONForm from "./components/DigitJSONForm";
 import LevelCards from "./components/LevelCards";
+import { Loader } from "@egovernments/digit-ui-components";
 
 import * as parsingUtils from "../src/utils/ParsingUtils"
 import CustomSwitch from "./components/CustomSwitch";
 
 const WorkbenchModule = ({ stateCode, userType, tenants }) => {
-  const moduleCode = ["workbench", "mdms", "schema", "hcm-admin-schemas"];
+  const moduleCode = ["workbench","mdms","schema","hcm-admin-schemas"];
   const match = useMatch("/employee/workbench/*");
   const path = match?.pathnameBase || "";
-
   const language = Digit.StoreData.getCurrentLanguage();
   const modulePrefix = window?.globalConfigs?.getConfig("CORE_UI_MODULE_LOCALE_PREFIX") || "rainmaker";
 
@@ -98,5 +96,5 @@ const updateCustomConfigs = () => {
   });
 };
 
-export   {initWorkbenchComponents, DigitJSONForm};
+export {initWorkbenchComponents, DigitJSONForm};