+++ micro-ui/web/micro-ui-internals/packages/modules/workbench/src/configs/LocalisationSearchConfig.js
@@ -83,7 +83,15 @@ export const Config = {
                 name: "codes",
             }
           },
-        
+          // {
+          //   label: "WBH_LOC_MESSAGE",
+          //   type: "text",
+          //   isMandatory: false,
+          //   disable: false,
+          //   populators: { 
+          //       name: "message",
+          //   }
+          // },
         ],
       },
       label: "",