+++ micro-ui/web/micro-ui-internals/packages/modules/workbench/src/configs/UICustomizations.js
@@ -469,6 +469,57 @@ export const UICustomizations = {
       data.body.MdmsCriteria.filters = filters
       data.body.MdmsCriteria.schemaCode = additionalDetails?.currentSchemaCode
       delete data.body.MdmsCriteria.custom
+      // const {field,value} = data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].custom || {}
+
+      // const tenantId = Digit.ULBService.getCurrentTenantId()
+      // data.body.MdmsCriteria.tenantId = tenantId
+
+      // //generate filter 
+      // const filter = `[?(@.${field?.code}=='${value}')]`
+
+
+      // data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].filter = filter
+      // delete data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].custom
+      //we simply
+      // data.params = { ...data.params, tenantId: Digit.ULBService.getCurrentTenantId() };
+
+      // let requestBody = { ...data.body.Individual };
+      // const pathConfig = {
+      //   name: "name.givenName",
+      // };
+      // const dateConfig = {
+      //   createdFrom: "daystart",
+      //   createdTo: "dayend",
+      // };
+      // const selectConfig = {
+      //   wardCode: "wardCode[0].code",
+      //   socialCategory: "socialCategory.code",
+      // };
+      // const textConfig = ["name", "individualId"];
+      // let Individual = Object.keys(requestBody)
+      //   .map((key) => {
+      //     if (selectConfig[key]) {
+      //       requestBody[key] = _.get(requestBody, selectConfig[key], null);
+      //     } else if (typeof requestBody[key] == "object") {
+      //       requestBody[key] = requestBody[key]?.code;
+      //     } else if (textConfig?.includes(key)) {
+      //       requestBody[key] = requestBody[key]?.trim();
+      //     }
+      //     return key;
+      //   })
+      //   .filter((key) => requestBody[key])
+      //   .reduce((acc, curr) => {
+      //     if (pathConfig[curr]) {
+      //       _.set(acc, pathConfig[curr], requestBody[curr]);
+      //     } else if (dateConfig[curr] && dateConfig[curr]?.includes("day")) {
+      //       _.set(acc, curr, Digit.Utils.date.convertDateToEpoch(requestBody[curr], dateConfig[curr]));
+      //     } else {
+      //       _.set(acc, curr, requestBody[curr]);
+      //     }
+      //     return acc;
+      //   }, {});
+
+      // data.body.Individual = { ...Individual };
       return data;
     },
     additionalCustomizations: (row, key, column, value, t, searchResult) => {
@@ -560,7 +611,57 @@ export const UICustomizations = {
       data.body.MdmsCriteria.isActive = true;
       data.body.MdmsCriteria.schemaCode = additionalDetails?.currentSchemaCode
       delete data.body.MdmsCriteria.custom
-      
+      // const {field,value} = data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].custom || {}
+
+      // const tenantId = Digit.ULBService.getCurrentTenantId()
+      // data.body.MdmsCriteria.tenantId = tenantId
+
+      // //generate filter 
+      // const filter = `[?(@.${field?.code}=='${value}')]`
+
+
+      // data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].filter = filter
+      // delete data.body.MdmsCriteria.moduleDetails[0].masterDetails[0].custom
+      //we simply
+      // data.params = { ...data.params, tenantId: Digit.ULBService.getCurrentTenantId() };
+
+      // let requestBody = { ...data.body.Individual };
+      // const pathConfig = {
+      //   name: "name.givenName",
+      // };
+      // const dateConfig = {
+      //   createdFrom: "daystart",
+      //   createdTo: "dayend",
+      // };
+      // const selectConfig = {
+      //   wardCode: "wardCode[0].code",
+      //   socialCategory: "socialCategory.code",
+      // };
+      // const textConfig = ["name", "individualId"];
+      // let Individual = Object.keys(requestBody)
+      //   .map((key) => {
+      //     if (selectConfig[key]) {
+      //       requestBody[key] = _.get(requestBody, selectConfig[key], null);
+      //     } else if (typeof requestBody[key] == "object") {
+      //       requestBody[key] = requestBody[key]?.code;
+      //     } else if (textConfig?.includes(key)) {
+      //       requestBody[key] = requestBody[key]?.trim();
+      //     }
+      //     return key;
+      //   })
+      //   .filter((key) => requestBody[key])
+      //   .reduce((acc, curr) => {
+      //     if (pathConfig[curr]) {
+      //       _.set(acc, pathConfig[curr], requestBody[curr]);
+      //     } else if (dateConfig[curr] && dateConfig[curr]?.includes("day")) {
+      //       _.set(acc, curr, Digit.Utils.date.convertDateToEpoch(requestBody[curr], dateConfig[curr]));
+      //     } else {
+      //       _.set(acc, curr, requestBody[curr]);
+      //     }
+      //     return acc;
+      //   }, {});
+
+      // data.body.Individual = { ...Individual };
       return data;
     },
     additionalCustomizations: (row, key, column, value, t, searchResult) => {
@@ -812,7 +913,7 @@ export const UICustomizations = {
       //action===EDIT go to edit screen
       else if (actionSelected === "EDIT") {
         const additionalParamString = new URLSearchParams(props?.additionalParams).toString();
-        props?.navigate(`/${window?.contextPath}/employee/workbench/mdms-edit?moduleName=${props?.moduleName}&masterName=${props?.masterName}&uniqueIdentifier=${props?.uniqueIdentifier}${additionalParamString ? "&" + additionalParamString : ""}`)
+        props?.history.push(`/${window?.contextPath}/employee/workbench/mdms-edit?moduleName=${props?.moduleName}&masterName=${props?.masterName}&uniqueIdentifier=${props?.uniqueIdentifier}${additionalParamString ? "&" + additionalParamString : ""}`)
       }
       //action===DISABLE || ENABLE call update api and show toast respectively
       else {